changed multer to firebase, added delete category when deleting post, preventing to create  duplicated category, added apicalls and so on

+++ api/routes/posts.js
@@ -45,12 +45,18 @@ router.put("/:id", async (req, res) => {
 router.delete("/:id", async (req, res) => {
     try {   
         const post = await Post.findById(req.params.id);
-        const category = await Category.findOne({ name: post.category });
     
         if (post.username === req.body.username) {
             try {
-                await post.delete();
-                await category.delete();
+                // if all posts has only 1 category, delete it with post                
+                const allPosts = await Post.find();
+                const category = await Category.findOne({ name: post.category });
+                const allCats = allPosts.filter(post => post.category == category.name);
+                if (allCats.length == 1) {
+                    await category.delete();
+                }
+
+                await post.delete();                
                 res.status(200).json("Post has been deleted.");
             } catch (err) {
                 res.status(500).json(err);
